refactor: view , page ,style 重構目錄
